Receiver is now mapped, and arduino sketch resends button well.
The receiver script is now mapped to a useful set of commands, and the
arduino sketch will resend button presses after ~600 ms.

Overall, it works well, however there is a large flaw: The arduino is
the one that is advertising and the receiver must connect to it. In
other words, the remote cannot reconnect to the receiver after turning
off and back on again.

One way to fix this is to have the receiver periodically reconnect.
Another way is to have the receiver advertise and the controller
connect.