Added heartbeat to receiver script.
The receiver script will restart if the connection is lost or gatttool
times out waiting for a connection. It is safer to have my laptop do a
heartbeat than advertise looking for a controller. This way, I know only
a BLE with the same MAC is connecting to my laptop (and not some random
BLE).